+++ src/modules/routing/components/TSR/TsrItem.vue
@@ -3,6 +3,7 @@ import {mapGetters, mapActions, mapMutations} from "vuex";
 import IconButton from "@shared/modules/buttons/components/IconButton.vue";
 import RoutingCoordinateInput from "../RoutingCoordinateInput.vue";
 import RoutingSpeedProfileIcon from "../RoutingSpeedProfileIcon.vue";
+import RoutingContextMenu from "../RoutingContextMenu.vue";
 import * as constantsRouting from "../../store/constantsRouting";
 import TsrUpload from "./TsrUpload.vue";
 import TsrOutput from "./TsrOutput.vue";
@@ -20,6 +21,7 @@ export default {
         IconButton,
         RoutingCoordinateInput,
         RoutingSpeedProfileIcon,
+        RoutingContextMenu,
         TsrUpload,
         TsrOutput
     },
@@ -311,6 +313,7 @@ export default {
             </div>
         </div>
     </div>
+    <RoutingContextMenu />
 </template>