+++ src/modules/routing/js/map/directions/waypoints/directionsWaypointsDraw.js
@@ -5,7 +5,7 @@ import directionsWaypointsSource from "./directionsWaypointsSource";
 export default new Draw({
     source: directionsWaypointsSource,
     type: "Point",
-    style: directionsWaypointsStyle
+    style: directionsWaypointsStyle,
     // allow only left click
-    // condition: e => e.pointerEvent.buttons === 1
+    condition: e => e.originalEvent.buttons === 1
 });