+++ src/modules/routing/tests/unit/components/Directions/DirectionsItem.spec.js
@@ -8,6 +8,7 @@ import RoutingBatchProcessingCheckboxComponent from "@modules/routing/components
 import RoutingDownloadComponent from "@modules/routing/components/RoutingDownload.vue";
 import RoutingRestrictionsInputComponent from "@modules/routing/components/RoutingRestrictionsInput.vue";
 import RoutingElevationProfileComponent from "@modules/routing/components/RoutingElevationProfile.vue";
+import RoutingContextMenuComponent from "../../../../components/RoutingContextMenu.vue";
 
 config.global.mocks.$t = key => key;
 
@@ -301,6 +302,17 @@ describe("src/modules/routing/components/Directions/DirectionsItem.vue", () => {
         await wrapper.vm.$nextTick();
         expect(wrapper.findComponent(RoutingElevationProfileComponent).exists()).to.be.false;
     });
+
+    it("renders context menu", async () => {
+        batchProcessingEnabled = false;
+        wrapper = shallowMount(DirectionsComponent, {global: {
+            plugins: [store]
+        }});
+
+        await wrapper.vm.$nextTick();
+        expect(wrapper.findComponent(RoutingContextMenuComponent).exists()).to.be.true;
+    });
+
 
     describe("checks if findDirections are called", () => {
         it("should call on changeSpeedProfile", () => {